refactor: expose tokio::main entrypoint in crate::main
