+++ Cargo.toml
@@ -3,7 +3,8 @@ name = "proxy-solver-api"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+rocket = []
 
 [dependencies]
 aws_lambda_events = "0.7.3"
