+++ src/main.rs
@@ -49,6 +49,7 @@ fn post_solution(log: Slogger, problem: Json<ProxyProblem>) -> Json<ProxySolutio
 }
 
 #[launch]
+#[cfg(feature = "rocket")]
 /// Sets up Rocket routing, plus extras for logging and OpenAPI generation.
 fn rocket() -> _ {
     let fairing = Slogger::new_bunyan_logger(env!("CARGO_PKG_NAME"));
@@ -59,6 +60,44 @@ fn rocket() -> _ {
     rocket::build()
         .attach(fairing)
         .mount("/", openapi_get_routes![get_health_ready, post_solution])
+}
+
+use aws_lambda_events::encodings::Body;
+use aws_lambda_events::event::apigw::{ApiGatewayProxyRequest, ApiGatewayProxyResponse};
+use http::header::HeaderMap;
+use lambda_runtime::{handler_fn, Context, Error};
+use log::LevelFilter;
+use simple_logger::SimpleLogger;
+
+#[tokio::main]
+#[cfg(not(feature = "rocket"))]
+async fn main() -> Result<(), Error> {
+    SimpleLogger::new()
+        .with_level(LevelFilter::Info)
+        .init()
+        .unwrap();
+
+    let func = handler_fn(my_handler);
+    lambda_runtime::run(func).await?;
+    Ok(())
+}
+
+#[cfg(not(feature = "rocket"))]
+pub(crate) async fn my_handler(
+    event: ApiGatewayProxyRequest,
+    _ctx: Context,
+) -> Result<ApiGatewayProxyResponse, Error> {
+    let path = event.path.unwrap();
+
+    let resp = ApiGatewayProxyResponse {
+        status_code: 200,
+        headers: HeaderMap::new(),
+        multi_value_headers: HeaderMap::new(),
+        body: Some(Body::Text(format!("Hello from '{}'", path))),
+        is_base64_encoded: Some(false),
+    };
+
+    Ok(resp)
 }
 
 #[cfg(test)]
